Edited font to Inter and created helper requests file
