+++ app/data.js
@@ -9,3 +9,16 @@ export const navData = [
     ]
   }
 ];
+
+export const nasdaq100 = [
+  'Apple Inc', 'Microsoft Corp', 'NVIDIA Corp', 'Amazon.com Inc', 'Broadcom Inc', 'Meta Platforms Inc', 'Costco Wholesale Corp', 'Netflix Inc', 'Tesla Inc', 'Alphabet Inc', 'T-Mobile US Inc', 'Cisco Systems Inc', 
+  'PepsiCo Inc', 'Palantir Technologies Inc', 'Intuitive Surgical Inc', 'Amgen Inc', 'Intuit Inc', 'Adobe Inc',
+  'Qualcomm Inc', 'Booking Holdings Inc', 'Advanced Micro Devices Inc', 'Texas Instruments Inc', 'Gilead Sciences Inc', 'Comcast Corp', 'Honeywell International Inc', 'Vertex Pharmaceuticals Inc', 'Automatic Data Processing Inc', 'Applied Materials Inc',
+  'Palo Alto Networks Inc', 'MercadoLibre Inc', 'Starbucks Corp', 'Intel Corp', 'Mondelez International Inc', 'Analog Devices Inc',
+  'Cintas Corp', 'KLA Corp', 'Lam Research Corp', 'CrowdStrike Holdings Inc', 'Micron Technology Inc', 'Microstrategy Inc', 'PDD Holdings Inc', 'Applovin Corp', 'Fortinet Inc', 'DoorDash Inc', 'Cadence Design Systems Inc', 'Regeneron Pharmaceuticals Inc', 
+  'Synopsys Inc', 'Marriott International Inc', 'Roper Technologies Inc', 'PayPal Holdings Inc', 'American Electric Power Company Inc', 'Monster Beverage Corp', 'ASML Holding NV', 'Autodesk Inc', 'Copart Inc', 'Paychex Inc',
+  'CSX Corp', 'Charter Communications Inc', 'Paccar Inc', 'Workday Inc', 'Airbnb Inc', 'Keurig Dr Pepper Inc', 'Exelon Corp', 'Ross Stores Inc', 'Marvell Technology Inc', 'Fastenal Co',
+  'NXP Semiconductors NV', 'Verisk Analytics Inc', 'AstraZeneca PLC', 'Xcel Energy Inc', 'Coca-Cola Europacific Partners PLC', 'Axon Enterprise Inc', 'Diamondback Energy Inc', 'Kraft Heinz Co', 'Electronic Arts Inc', 'Baker Hughes Co',
+  'Take-Two Interactive Software Inc', 'Cognizant Technology Solutions Corp', 'Old Dominion Freight Line Inc', 'IDEXX Laboratories Inc', 'Atlassian Corporation Ltd', 'Lululemon Athletica Inc', 'CoStar Group Inc', 'Datadog Inc', 'Zscaler Inc',
+  'ANSYS Inc', 'Dexcom Inc', 'Trade Desk Inc', 'Warner Bros Discovery Inc', 'Microchip Technology Inc', 'CDW Corp', 'Biogen Inc', 'ON Semiconductor Corp', 'MongoDB Inc'
+]
